booking-platform/requirements/requirements.txt
# --------------------------------------------------------
# WARNING: Do not edit the contents of this file directly;
# You should modify base.txt and then run freeze.sh
# --------------------------------------------------------
aiohttp==3.8.5
amqp==5.1.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
boto3==1.28.6
botocore==1.31.8
CacheControl==0.13.1
cachetools==5.3.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
chardet==5.1.0
createsend==7.0.0
cssutils==2.7.1
dacite==1.8.1
dj-database-url==2.0.0
Django==4.2.3
django-allianceutils==3.0.0
django-cors-headers==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-hijack==3.4.1
# https://github.com/roverdotcom/django-inlinecss/issues/62
django-inlinecss@git+https://github.com/blocher/django-inlinecss.git@django4
#django-inlinecss==0.3.0
django-mssql-backend==2.8.1 ; "aws" in platform_release or sys_platform == 'darwin'
django-ordered-model==3.7.4
django-storages==1.13.2
-e git+https://github.com/levic/django-webpack-loader.git@b17e7d3420250cedc46bcf9e44ea7bd2ae1cf166#egg=django_webpack_loader
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
-e git+https://gitlab.alliancesoftware.com.au/alliance/djrad.git@4d40fb2cad1f11120e35397477314591856354b9#egg=djrad&subdirectory=packages/djrad
-e git+https://gitlab.alliancesoftware.com.au/alliance/djrad.git@4d40fb2cad1f11120e35397477314591856354b9#egg=djrad_devtools&subdirectory=packages/djrad_devtools
-e git+https://gitlab.alliancesoftware.com.au/alliance/djrad.git@4d40fb2cad1f11120e35397477314591856354b9#egg=djrad_rest&subdirectory=packages/djrad_rest
drf-writable-nested==0.7.0
escapejson==1.0
-e git+https://github.com/levic/factory_boy.git@47debe87c36e15aa84b5ada25e9cb92948a246d7#egg=factory_boy
Faker==19.2.0
fastprogress==1.0.3
firebase-admin==6.2.0
freezegun==1.2.2
future==0.18.3
google-api-core==2.11.1
google-api-python-client==2.94.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.1
googlemaps==4.10.0
grpcio==1.56.0
gunicorn==21.2.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
jmespath==1.0.1
kombu==5.3.1
msgpack==1.0.5
multidict==6.0.4
numpy==1.25.1
packaging==23.1
pdfkit==1.0.0
Pillow==10.0.0
protobuf==4.23.4
psycopg2==2.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pynliner==0.8.0
pyodbc==4.0.39 ; "aws" in platform_release or sys_platform == 'darwin'
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz
pywatchman==1.4.1
redis==4.6.0
requests==2.31.0
rsa==4.9
rules==3.3
s3transfer==0.6.1
sentry-sdk==1.28.1
setuptools==65.5.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
text-unidecode==1.3
typing-extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
vine==5.0.0
webencodings==0.5.1
yarl==1.9.2
zipp==3.16.2
